app1_notes/requirements.txt
flask>=3.0
flask-login>=0.6
flask-sqlalchemy>=3.1
anthropic>=0.25
face_recognition>=1.3
Pillow>=10.0
python-dotenv>=1.0
